+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4] - Unreleased
+
+### Added
+
+- Add ``row_wise_max`` function
+
+### Changed
+
+- Refactor ``row_wise_sum`` function to ``misc`` module
+- Change ``row_wise_sum`` to not make use of ``torch.autograd.Function`` 
+
 ## [1.3] - 2024-09-26
 
 ### Added